refactor(bin-designer): split svgParser into focused modules
Decompose the 855-LOC svgParser.ts into 5 sibling files organized by
concern. The parser entry function and ViewBox type stay in svgParser.ts
(now 153 LOC); everything else moves to siblings.

Extracted modules (all under 260 LOC):

- svgTransform.ts (134) — Matrix tuple + identity, multiplyMatrices,
  applyMatrix, parseTransformFunction (translate/scale/rotate/matrix/
  skewX/Y), parseTransformAttr, resolveTransformChain (walks ancestor
  chain), transformPoint (matrix + viewBox + Y-flip combined),
  isIdentityOrTranslate.
- svgArcToBezier.ts (152) — BezierSegment interface + the W3C-spec
  arc → cubic bezier converter (arcToCubicBeziers), plus the small
  vectorAngle and arcSegmentToBezier helpers.
- svgConvertShapes.ts (257) — primitive-shape converters
  (convertRect / Circle / Ellipse / PointsElement) plus the helpers
  shared with path conversion: wrapSingle, numAttr, flipY,
  makeCornerPoint, parsePointsAttr, pointsToPathSpec, pathPointsToSpec,
  convertCircleAsPath. Each primitive uses a fast path when the matrix
  is identity-or-translate; complex transforms rasterize through
  pointsToPathSpec / convertCircleAsPath.
- svgConvertPath.ts (218) — convertPath (top-level <path>) +
  splitContours (each M starts a sub-path) + convertContour (the big
  command-stream reducer).

svgParser.ts: 855 LOC -> 153 LOC. The public surface (parseSvgString +
ViewBox type) is unchanged; consumers compile without changes.

3 svg test files / 39 tests pass (916ms targeted run).
